+++ internal/filepublication/publication_darwin.go
@@ -1,0 +1,9 @@
+//go:build darwin
+
+package filepublication
+
+import "golang.org/x/sys/unix"
+
+func publishNoReplace(temporary, path string) error {
+	return unix.RenamexNp(temporary, path, unix.RENAME_EXCL)
+}